Have solver write solved board to file
The solver now writes a string representation
of the solved board to a file if the puzzle
was solvable.

Some more documentation and formatting has
been added to the notebook.

The README has been updated to account for
the new methods related to file writing.
Existing methods in the README have also been
updated to include their parameters and
expected return values.